requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   starlette
    #   watchfiles
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via uvicorn
fastapi==0.135.3
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httptools==0.7.1
    # via uvicorn
idna==3.11
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via -r requirements.in
lxml==6.0.4
    # via -r requirements.in
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
pillow==12.2.0
    # via -r requirements.in
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   uvicorn
python-magic==0.4.27
    # via -r requirements.in
python-multipart==0.0.26
    # via -r requirements.in
pyyaml==6.0.3
    # via uvicorn
requests==2.33.1
    # via -r requirements.in
starlette==1.0.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.44.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
werkzeug==3.1.8
    # via -r requirements.in
